requirements.txt
altgraph==0.16.1
future==0.17.1
pefile==2019.4.18
PyInstaller==3.5
pyserial==3.4
pywin32-ctypes==0.2.0
XlsxWriter==1.1.8
